Write C++ implementation for points checking
Finding lat/longs in range of points takes quadratic time which is slow. Even though the program is unlikely to have extreme input, this ensures around x10 performance especially due to faster haversine calculation. code adapted accordingly

+++ src/activity.py
@@ -79,7 +79,7 @@ class Activity:
     elevation: float | None
     elevation_per_km: float | None
     cadence: float | None
-    lat_long_stream: list[list[int]] | None
+    lat_long_stream: list[list[float]] | None
     heart_rate_stream: list[int] | None
     weather: Weather | None
